Simplify and fix dao channel logic switching
