+++ elements/chat/ChatAuthButton.tsx
@@ -2,41 +2,23 @@ import { useEffect } from "react";
 import { useAppDispatch, useAppSelector } from "../../hooks/redux";
 import { Button, Typography } from "../../components";
 import { usePhantom } from "../../hooks/usePhantom";
-import { signMessageWithKey } from "../../utils/signMessageWithKey";
 import { fetchChatUserToken } from "../../store/chatSlice";
 import styled from "styled-components/native";
+import { ActivityIndicator } from "react-native";
 
 export const ChatAuthButton = () => {
   const { signedMessage, signMessage } = usePhantom();
   const { publicKey, walletType } = useAppSelector((state) => state.wallet);
-  const { chatUserToken, isAuthenticating } = useAppSelector(
+  const dispatch = useAppDispatch();
+  const { isAuthenticating, chatUserToken } = useAppSelector(
     (state) => state.chat
   );
-  const { selectedRealm } = useAppSelector((state) => state.realms);
-
-  const dispatch = useAppDispatch();
 
   useEffect(() => {
     if (signedMessage) {
       dispatch(fetchChatUserToken({ publicKey, signedMessage: signedMessage }));
     }
   }, [signedMessage]);
-
-  useEffect(() => {
-    // If wallet type is web3auth we can automatically fetch user token for chat
-    const loginWithWeb3auth = async () => {
-      const message =
-        "Proving DAO membership to authenticate into Gilder Chat.";
-      const web3AuthSignedMessage = await signMessageWithKey(message);
-      dispatch(
-        fetchChatUserToken({ publicKey, signedMessage: web3AuthSignedMessage })
-      );
-    };
-
-    if (publicKey && walletType === "web3auth") {
-      loginWithWeb3auth();
-    }
-  }, [publicKey, selectedRealm?.pubKey]);
 
   const authIntoChat = async () => {
     if (walletType === "phantom") {
@@ -45,6 +27,10 @@ export const ChatAuthButton = () => {
       console.log("no wallet connected");
     }
   };
+
+  if (isAuthenticating) {
+    return <ActivityIndicator />;
+  }
 
   return (
     <SignInContainer>
